Corrected variable data type.

+++ src/main.cpp
@@ -40,7 +40,7 @@ float integral_min = 0.0f;
 float derivative = 0.0f;
 int interval = 1; //s
 int pwm_output_max = 100;
-float pwm_output_min = 0.0f;
+int pwm_output_min = 0;
 int pwm_output = 0;
 
 // Define the GPIO pin to use for PWM output
